+++ src/main/kotlin/lotto/calculation/TicketGenerator.kt
@@ -7,7 +7,7 @@ class TicketGenerator {
 
     // Function to generate a single ticket
     private fun generateSingleTicket(): Lotto {
-        val numbers = Randoms.pickUniqueNumbersInRange(1, 45, 6)
+        val numbers = Randoms.pickUniqueNumbersInRange(1, 45, 6).sorted()
 
         return Lotto(numbers)
     }